pyproject.toml
[tool.poetry]
name = "cerverus"
version = "0.1.0"
description = "Financial Anomaly Detection System"
authors = ["Your Name <your@email.com>"]
readme = "README.md"
packages = [{ include = "cerverus", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
# Core Framework
fastapi = "^0.109.1"
uvicorn = { extras = ["standard"], version = "^0.24.0" }
pydantic = "^2.5.0"
python-dotenv = "^1.0.0"
# Data Processing - SIN TensorFlow para build rápido
pandas = "^2.0.0"
numpy = "^1.24.0"
scikit-learn = "^1.3.0"
yfinance = "^0.2.21"
requests = "^2.31.0"
# Database
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.23"
# Utilities
python-multipart = "^0.0.18"
# Visualization (required by isolation_forest_eda.py)
matplotlib = "^3.7.0"
seaborn = "^0.12.0"
# Development
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
